+++ pytorch_lightning/utilities/cli.py
@@ -24,7 +24,8 @@ from pytorch_lightning.utilities.seed import seed_everything
 
 _JSONARGPARSE_AVAILABLE = _module_available("jsonargparse")
 if _JSONARGPARSE_AVAILABLE:
-    from jsonargparse import ActionConfigFile, ArgumentParser
+    from jsonargparse import ActionConfigFile, ArgumentParser, set_config_read_mode
+    set_config_read_mode(fsspec_enabled=True)
 else:
     ArgumentParser = object