+++ tests/trainer/test_trainer.py
@@ -624,9 +624,8 @@ def test_tested_checkpoint_path(tmpdir, ckpt_path, save_top_k, fn):
         def test_step(self, *args):
             return self.validation_step(*args)
 
-        def predict_step(self, *args):
-            args = args[:-1]  # remove `dataloader_idx`
-            return self.validation_step(*args)
+        def predict_step(self, batch, *_):
+            return self(batch)
 
     model = TestModel()
     model.test_epoch_end = None
@@ -1912,30 +1911,30 @@ def test_on_load_checkpoint_missing_callbacks(tmpdir):
 
 
 def test_module_current_fx_attributes_reset(tmpdir):
-    """ Ensure that lightning module's attributes related to current hook fx are reset at the end of execution. """
+    """ Ensure that lightning module's attributes related to current fx are reset at the end of execution. """
     model = BoringModel()
-    model.validation_step = None
-    model.training_epoch_end = None
     trainer = Trainer(
         default_root_dir=tmpdir,
-        max_epochs=1,
+        fast_dev_run=1,
         checkpoint_callback=False,
         logger=False,
-        limit_val_batches=0,
     )
+
     trainer.fit(model)
-    assert model._current_fx_name == "", f"_current_fx_name not reset after fit: {model._current_fx_name}"
-    assert (
-        model._current_hook_fx_name is None
-    ), f"_current_hook_fx_name not reset after fit: {model._current_hook_fx_name}"
-    assert (
-        model._current_dataloader_idx is None
-    ), f"_current_dataloader_idx not reset after fit: {model._current_dataloader_idx}"
+    assert model._current_fx_name is None
+    assert model._current_dataloader_idx is None
+
     trainer.test(model)
-    assert model._current_fx_name == "", f"_current_fx_name not reset after test: {model._current_fx_name}"
-    assert (
-        model._current_hook_fx_name is None
-    ), f"_current_hook_fx_name not reset after test: {model._current_hook_fx_name}"
-    assert (
-        model._current_dataloader_idx is None
-    ), f"_current_dataloader_idx not reset after test: {model._current_dataloader_idx}"
+    assert model._current_fx_name is None
+    assert model._current_dataloader_idx is None
+
+
+def test_exception_when_lightning_module_is_not_set_on_trainer():
+    trainer = Trainer()
+
+    with pytest.raises(MisconfigurationException, match=r"`model` must be provided.*validate"):
+        trainer.validate()
+    with pytest.raises(MisconfigurationException, match=r"`model` must be provided.*test"):
+        trainer.test()
+    with pytest.raises(MisconfigurationException, match=r"`model` must be provided.*predict"):
+        trainer.predict()